fix(docs): refactor policy_engine api and docstrings
