+++ src/macaron/policy_engine/examples/simple_example.dl
@@ -4,5 +4,5 @@
 
 #include "prelude.dl"
 
-Policy("auth-provenance", repositoryid, "") :-    check_passed(repositoryid, "mcn_provenance_level_three_1").
+Policy("auth-provenance", repositoryid, "") :- check_passed(repositoryid, "mcn_provenance_level_three_1").
 apply_policy_to("auth-provenance", repo) :- is_repo(repo, _).
